+++ cmd/zetaclientd/start.go
@@ -5,6 +5,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"net/http"
+	_ "net/http/pprof" // #nosec G108 -- pprof enablement is intentional
 	"os"
 	"os/signal"
 	"path/filepath"
@@ -231,6 +233,7 @@ func Start(_ *cobra.Command, _ []string) error {
 	go func() {
 		host := tssServer.GetP2PHost()
 		pingRTT := make(map[peer.ID]int64)
+		pingRTTLock := sync.Mutex{}
 		for {
 			var wg sync.WaitGroup
 			for _, p := range whitelistedPeers {
@@ -240,6 +243,8 @@ func Start(_ *cobra.Command, _ []string) error {
 					ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 					defer cancel()
 					result := <-ping.Ping(ctx, host, p)
+					pingRTTLock.Lock()
+					defer pingRTTLock.Unlock()
 					if result.Error != nil {
 						masterLogger.Error().Err(result.Error).Msg("ping error")
 						pingRTT[p] = -1 // RTT -1 indicate ping error
@@ -251,6 +256,15 @@ func Start(_ *cobra.Command, _ []string) error {
 			wg.Wait()
 			telemetryServer.SetPingRTT(pingRTT)
 			time.Sleep(30 * time.Second)
+		}
+	}()
+	// pprof http server
+	// zetacored/cometbft is already listening for pprof on 6060 (by default)
+	go func() {
+		// #nosec G114 -- timeouts uneeded
+		err := http.ListenAndServe("localhost:6061", nil)
+		if err != nil {
+			log.Error().Err(err).Msg("pprof http server error")
 		}
 	}()