+++ e2e/e2etests/e2etests.go
@@ -55,16 +55,17 @@ const (
 	/*
 	 * Solana tests
 	 */
-	TestSolanaDepositName                   = "solana_deposit"
-	TestSolanaWithdrawName                  = "solana_withdraw"
-	TestSolanaDepositAndCallName            = "solana_deposit_and_call"
-	TestSolanaDepositAndCallRefundName      = "solana_deposit_and_call_refund"
-	TestSolanaDepositRestrictedName         = "solana_deposit_restricted"
-	TestSolanaWithdrawRestrictedName        = "solana_withdraw_restricted"
-	TestSPLDepositName                      = "spl_deposit"
-	TestSPLDepositAndCallName               = "spl_deposit_and_call"
-	TestSPLWithdrawName                     = "spl_withdraw"
-	TestSPLWithdrawAndCreateReceiverAtaName = "spl_withdraw_and_create_receiver_ata"
+	TestSolanaDepositName                      = "solana_deposit"
+	TestSolanaWithdrawName                     = "solana_withdraw"
+	TestSolanaDepositAndCallName               = "solana_deposit_and_call"
+	TestSolanaDepositAndCallRevertName         = "solana_deposit_and_call_revert"
+	TestSolanaDepositAndCallRevertWithDustName = "solana_deposit_and_call_revert_with_dust"
+	TestSolanaDepositRestrictedName            = "solana_deposit_restricted"
+	TestSolanaWithdrawRestrictedName           = "solana_withdraw_restricted"
+	TestSPLDepositName                         = "spl_deposit"
+	TestSPLDepositAndCallName                  = "spl_deposit_and_call"
+	TestSPLWithdrawName                        = "spl_withdraw"
+	TestSPLWithdrawAndCreateReceiverAtaName    = "spl_withdraw_and_create_receiver_ata"
 
 	/**
 	 * TON tests
@@ -109,9 +110,10 @@ const (
 	 Miscellaneous tests
 	 Test various functionalities not related to assets
 	*/
-	TestContextUpgradeName = "context_upgrade"
-	TestMyTestName         = "my_test"
-	TestDonationEtherName  = "donation_ether"
+	TestContextUpgradeName  = "context_upgrade"
+	TestMyTestName          = "my_test"
+	TestDonationEtherName   = "donation_ether"
+	TestInboundTrackersName = "inbound_trackers"
 
 	/*
 	 Stress tests
@@ -170,6 +172,7 @@ const (
 	TestV2ZEVMToEVMCallName                      = "v2_zevm_to_evm_call"
 	TestV2ZEVMToEVMCallThroughContractName       = "v2_zevm_to_evm_call_through_contract"
 	TestV2EVMToZEVMCallName                      = "v2_evm_to_zevm_call"
+	TestV2DepositAndCallSwapName                 = "v2_deposit_and_call_swap"
 
 	/*
 	 Operational tests
@@ -453,12 +456,18 @@ var AllE2ETests = []runner.E2ETest{
 		TestSPLWithdrawAndCreateReceiverAta,
 	),
 	runner.NewE2ETest(
-		TestSolanaDepositAndCallRefundName,
-		"deposit SOL into ZEVM and call a contract that reverts; should refund",
+		TestSolanaDepositAndCallRevertName,
+		"deposit SOL into ZEVM and call a contract that reverts",
 		[]runner.ArgDefinition{
 			{Description: "amount in lamport", DefaultValue: "1200000"},
 		},
-		TestSolanaDepositAndCallRefund,
+		TestSolanaDepositAndCallRevert,
+	),
+	runner.NewE2ETest(
+		TestSolanaDepositAndCallRevertWithDustName,
+		"deposit SOL into ZEVM; revert with dust amount that aborts the CCTX",
+		[]runner.ArgDefinition{},
+		TestSolanaDepositAndCallRevertWithDust,
 	),
 	runner.NewE2ETest(
 		TestSolanaDepositRestrictedName,
@@ -729,6 +738,12 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in wei", DefaultValue: "100000000000000000"},
 		},
 		TestDonationEther,
+	),
+	runner.NewE2ETest(
+		TestInboundTrackersName,
+		"test processing inbound trackers for observation",
+		[]runner.ArgDefinition{},
+		TestInboundTrackers,
 	),
 	/*
 	 Stress tests
@@ -1058,6 +1073,12 @@ var AllE2ETests = []runner.E2ETest{
 		"evm -> zevm call using V2 contract",
 		[]runner.ArgDefinition{},
 		TestV2EVMToZEVMCall,
+	),
+	runner.NewE2ETest(
+		TestV2DepositAndCallSwapName,
+		"evm -> zevm deposit and call with swap and withdraw back to evm",
+		[]runner.ArgDefinition{},
+		TestV2DepositAndCallSwap,
 	),
 	/*
 	 Special tests